fix: the base simulator got refactored
