--- Runtime/Interaction/BaseTeleportationSimulator.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 52afb6f360a14a228ba19efb52d445da
-timeCreated: 1617141296
