--- Runtime/Interaction/TeleportationSimulator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e2c9dac9f252a437f836fbd496b845c6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
